backend/pyproject.toml
[project]
name = "commai-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.119.0",
  "uvicorn[standard]==0.37.0",
  "starlette==0.48.0",
  "pydantic==2.12.0",
  "python-dotenv>=1.0.1",
  "httpx==0.28.1",
  "openai==2.3.0",
  "anthropic==0.69.0",
  "google-generativeai==0.8.5",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
